requirements.txt
appdirs==1.4.3
Django==1.11
django-filter==1.0.2
djangorestframework==3.6.2
Markdown==2.6.8
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
six==1.10.0
